Given two strings s and t, return the number of distinct subsequences of s which equals t.

+++ 115_Distinct_Subsequences.cpp
@@ -6,20 +6,20 @@ using namespace std;
 class Solution {
 public:
     int numDistinct(string s, string t) {
-          int m = s.size();
+        int m = s.size();
         int n = t.size();
-        
-        vector<long long> dp(n + 1, 0);
+
+        vector<unsigned long long> dp(n + 1, 0);
         dp[0] = 1;
-        
-        for(int i = 1; i <= m; i++) {
-            for(int j = n; j >= 1; j--) {
-                if(s[i - 1] == t[j - 1]) {
+
+        for (int i = 1; i <= m; i++) {
+            for (int j = n; j >= 1; j--) {
+                if (s[i - 1] == t[j - 1]) {
                     dp[j] += dp[j - 1];
                 }
             }
         }
-        
-        return dp[n];
+
+        return (int)dp[n];
     }
 };
